Fix some missed `${var}` deprecations (#41188)
Two in packages/analyzer, which disables all phpcompatibility sniffs.

One in a file in jetpack-mu-wpcom which has `phpcs:disable` at the top.

And one in a comment in plugins/beta, because why not.

The rest I found while grepping seem to be inside single-quoted strings
or in embedded JS code.