renovate: Work around bug in `atomically` package (#41261)
When #40842 updated `configstore`, that brought in `atomically` v2.0.3
which has a bug where it crashes if the current user is not listed in
`/etc/passwd`, which happens inside our docker container during the
renovate run.

Work around this by making sure an appropriate entry is added to
`/etc/passwd` inside the container.