RDV: Fix a performance issue and assign a12s treatment (#41378)
* RDV: Fix a performance issue and assign a12s treatment

Fixed an issue that caused to make an API request to ExPlat on each page request for users that were in control or if they were a12s.

This PR also forces treatment to all a12s on both Simple and Atomic sites.

On Simple sites, we use is_automattician since that's more accurate than relying on proxy. Although not needed for performance, the result is also cached in the user option so that a12s to be able to use the escape hatch.

On Atomic sites, the logic is a bit more complicated. The switch for a12s is handled with two conditions: Calypo: if there's an admin_menu_a11n query param (which is passed by WPCOM admin-menu API mapper

WP-Admin: Rely on the is proxy check.

Since this PR comes after a revert, we also need to change the caching key to pull again the experiment assignment from ExPlat.

* Restrict the parameter to the admin-menu endpoint and fix a linting issue