Phan: Fix errant extra params in function calls (#41263)
* Remove extra param on remove_filter/remove_action

There's no arg count when removing.

* Remove extra param from rest_ensure_response

That function doesn't allow a status code to be passed. Presumably we could override the WP_REST_Response status and force 200, but passing it through has worked so far, so keeping it that way.

* Remove other stray args

In most cases, these appear to be due to refactors.

* Another stray arg

* Suppress things that confuse Phan

* Fix get_avatar_url calls

* arrayHasKey → assertArrayHasKey

* Fix misplaced parentheses

* Tighten types

We're using `WP_Option`, which allows an extra param in its `get()`.

* Add missing sprintf

* Add unused param to allow dynamic method call

* Add changelogs

* Use same behaviour as trunk

* Fix assertion

* Add param to Storage::get and use that type as before

* Inform Phan of type rather than suppressing error

* Update errant types

* Update Phan baselines

* More changelogs