tools: Fix php-test-env composer.lock issue (#41817)
With the way things are currently set up, changes to
`tools/php-test-env/composer.json` will cause problems for existing
installs because `jetpack install -r` will try to reuse the untracked
`tools/php-test-env/composer.lock` from a previous install and fail
because it's not up to date with the composer.json.

There are a few ways we might fix this:

1. Disable composer.lock for that dir.
2. Check in composer.lock for that dir.
3. Have the post-install-cmd do `composer update`.

I decided to go for option 1 here, since there doesn't seem to be much
need to lock stuff and wait for the Renovate lockfile update job to get
any updates; previous to the introduction of php-test-env, we weren't
locking the deps for any of the package tests, so why change it?

I also took the opportunity to remove some unused deps from
`tools/php-test-env/composer.json` and bump the php version to match the
rest of the monorepo.