Boost: Introduce base minify class for JS and CSS modules (#41716)
* Add function that unschedules the 404 tester

* changelog

* Clear the cache cleanup cron too

* Clear the 404 tester cron on plugin deactivation too

* Schedule the daily jobs from the setup, rather than activation

If the plugin is deactivated with the concatenation modules enabled
then the activate function isn't called when the plugin is activated
again.
The setup function runs on every pageload, but it does an is_admin()
check to reduce DB queries.

* Revert changes to setup function

* Use base Minify class for both modules

* This code doesn't belong here

* Revert previous changes and rename Minify php file

* Track how many minify modules are active in Minify

* Changelog

* Remove duplicate deactivation function

* Call Minify activate and deactivate from JS and CSS modules

* Move these functions back to where they were before

* move get_slug around too

* Activate the Minify module by mapping js/css to it

* Schedule before testing as scheduling fails.

I'm not sure why, but if I schedule the job before running the test, the
schedule fails.

* No need for this, it's called by activation.

The 404 tester is called when it's scheduled

* Schedule the tester, and let cron run the first test immediately

* Module must be "activated" to call the on_activate function

Without this change, the on_activate() function would be called when one
module was toggled off, but the other was on, since the previous
! jetpack_boost_minify_is_enabled() check would be false.
We only want on_deactivate() to be called when both are disabled, but
on_activate() must be called whenever either module is activated.

* And revert back again. Need the tester for the UI

* Restore this blank line.

* Update projects/plugins/boost/app/lib/Status.php

Use the get_slug() method to return the slug.

Co-authored-by: Liam Sarsfield <43409125+LiamSarsfield@users.noreply.github.com>

* Fix fatal errors from last code checked in.

* Use sub modules for the Minify module

This will allow this code to remain generic and still deactivate the
concatenation modules correctly

* Make sure it's an array before using the sub modules.

* Call the tester after, so it doesn't break scheduling

* Use get_available_submodules() to get the submodules

* Remove the activate functions, they're not used here

---------

Co-authored-by: Donncha O Caoimh <5656673+donnchawp@users.noreply.github.com>
Co-authored-by: Liam Sarsfield <43409125+LiamSarsfield@users.noreply.github.com>