changelogger, stub-generator: Simplify bin by dropping Composer <2.2 support (#41876)
Composer 2.2 added a global `$_composer_autoload_path` that points to
the autoloader. We can simplify the bin files significantly by requiring
this Composer version and using that constant instead of trying to guess
where the autoloader lives.

This also requires various tools be updated to run changelogger and
stub-generator via the symlink added in `vendor/bin/` rather than trying
to use `bin/` directly.