codesniffer: Add Jetpack.PHPUnit.TestClassName sniff (#41930)
PHPUnit 10 requires that the test class name match the filename, PSR-4
style. It also recommends that test class names end in "Test". This new
sniff checks for both of these.

Our sniff also wants custom TestCase base classes to end in "TestCase"
or "TestBase" so it can recognize tests. The new sniff assumes any
abstract TestCase subclass is one of these and looks for that class
naming too.

Also, we're already on 7.4 because of dependencies, may as well bump to
8.0 so we can use `str_ends_with()`.