ci: Don't try to run Tests / Publish coverage data on forked PRs (#42229)
`pull_request` runs on PRs from forks don't have access to secrets, so
the job will fail. Better to not run it at all.

Best would be to figure out some way to run just this one job in the
proper context where it has access to the needed secrets, but that will
be more difficult. Let's leave it for a followup.