Repo Gardening: do not add label to assigned issues (#42246)
Fixes https://github.com/Automattic/studio/issues/1012

Until now, when a PR was opened to close a specific issue, the assignIssues task did 2 things:

1. Assign the issue to the person working on the PR, to indicate to everyone following the issue that it is being worked on, and to hold the person opening the PR accountable.
2. Add the "[Status] In Progress" label to indicate that the issue is being worked on.

With the GitHub UI improvements in the past few years, PRs opened to close a specific issue are a lot more visible than they once were. The "In Progress" label is consequently not that useful anymore. Let's remove it.