Customizer: Remove Customizer link from admin menu (#42397)
* Customizer: Remove Customizer link from admin menu

We are now hooking the logic that removes the customizer depending on the execution context (API vs regular request).

This is needed because  when executed on init, on API requests, the switch to blog is not executed on time, thus `wp_is_block_theme` returns the placeholder theme.

This PR also removes some duplicate code that's no longer necessary.

* Remove another function

* Remove the Custom CSS customizer