Launchpad: Do not hide the site title task from the updated write launchpad (#42410)
* Do not hide the site title task from the updated write launchpad

The site title task was hidden for the write intent in #33291 because
the write onboarding flow already included steps for setting the site
title. This is no longer the case. So it shouldn't be hidden on any
launchpad (i.e. it isn't hidden depending on `launchpad_context`).

However it should still be hidden on launched sites. After the blog
is deployed, My Home sees another checklist with a different checklist
slug. We don't want to make a change where the site title appears there.

* changelog