Add subscribers in wp-admin starting page (#42066)
This commit replaces the subscriber's link to Jetpack Cloud with a starting page for subscribers rendered with React.
The idea is to make things easier to review. This PR just contains what is required to have a menu item render a page with react.
The replacement of the external link with the new wp-admin-based page just happens when the jetpack_wp_admin_subscriber_management_enabled feature flag is active.