Fix Search Results Count Not Being Returned (#42570)
* Fix search result count

This commit fixes a bug with how search results are displayed. Currently,
combinging hasQuery with hasCorrectedQuery causes the if that was on line
70 to return false, which triggers the default return of the function, around
line 118 with "Showing popular results" (which should just be a default state
on first load, etc.).

This adds "smarter" logic handling so that we show the correct "Founx X result/s"

* changelog

* Update fix-search-results-count

* Move hasQuery check below isMultiSite check

- Move the hasQuery check below the isMultiSite check so that we run the latter
first, otherwise, it will never be reached if hasQuery is first.

- Minor update to `_n()` output when there is no corrected query.

* Update fix-search-results-count

* Update changelog message