Publicize: Fix the Jetpack plugin version check (#42649)
* Publicize: Fix the Jetpack version check

In order to know whether to load the new package versions of the editor
sidebar code, we check the Jetpack version. Previously we were checking
for less than or equal to 14.4-a.7, but it seems that the change didn't
make the 14.4 cut at all, and there has been a point release 14.4.1, so
it seems like that should contain the new code, when it doesn't.

This updates the version check to 14.5-a.1, which was the first release
with the new Publicize package.

* changelog