Newsletter Widget: improve the display of links. (#42645)
* Add changelog

* Newsletter Widget: ensure external link icons have proper margin

The icon styles are not brought over to wp-admin yet, so we need to bring them with the Newsletter widget styles.

* Adjust wording: remove dash and break into 2 sentences

* Newsletter Widget: only display External Icon when necessary

On WordPress.com simple sites, folks are always connected to both wp-admin and WordPress.com (Calypso), so we do not have to show them an external link ; they will remain on the platform.

* Remove unnecessary React import

* Create new component to render link depending on the platform

* Swich to using new component, and extract settings link

* Add more tests

* Bring in tracking for links, from #42660

Co-authored-by: Chris Holder <chris.holder@automattic.com>

* Better wording for Dashboard link component prop

We do not necessarily always want to link out when the site is not WordPress.com.

* Switch some of the links to always internal

* Update tests to support tracking in extracted link component

* Extract new component to its own file, with own test

See https://github.com/Automattic/jetpack/pull/42645#pullrequestreview-2718661382

---------

Co-authored-by: Chris Holder <chris.holder@automattic.com>