feat: Block editor assets endpoint (#42314)
* feat: Add block editor assets endpoint

The REST API endpoint returns all the styles and scripts to be loaded
for the post editor. This is useful when recreating the `post-new.php`
editor environment in a different context--e.g., a WebView within a
mobile app. Access to the assets allows loading plugin-provided blocks.

* feat: Match WP REST API editor assets paths

In additional to WordPress.com API URLs, we should match WP REST API
URLs as well. This is used for Atomic and self-hosted sites.

* refactor: Relocate block editor assets endpoint to Jetpack

We plan to utilize this endpoint for all site types, so it makes sense
to place it within Jetpack now.

* changelog

* test: Add initial block editor assets endpoint tests

* style: Address lint warnings

* test: Attempt fix for CI-only test failure

The test fails in the "PHP tests: PHP 8.1 WP latest with wpcomsh" CI
test run with the following:

```
fopen(/tmp/wordpress-latest/src/wp-content/mu-plugins/wpcomsh/jetpack_vendor/automattic/jetpack-mu-wpcom/src/build/block-inserter-modifications/block-inserter-modifications.asset.php): Failed to open stream: No such file or directory
```

* test: Attempt fix for CI-only test failure (2)

The test fails in the "PHP tests: PHP 8.1 WP latest with wpcomsh" CI
test run with the following:

```
fopen(/tmp/wordpress-latest/src/wp-content/mu-plugins/wpcomsh/jetpack_vendor/automattic/jetpack-mu-wpcom/src/build/block-inserter-modifications/block-inserter-modifications.asset.php): Failed to open stream: No such file or directory
```

* test: Fix missing build asset errors

Not all assets are built for the CI environment, so we need to dequeue
those assets in the test to avoid errors.

* test: Refactor test to align with existing test practices

* feat: Remove manual loading of WP Admin resources

This is unneeded at this time, as the originally targeted Jetpack
resource (`jetpack-connect`) is no longer conditionally registered.

* docs: Update documentation for Jetpack

* test: Explore simplified assets queue

Limit the enqueued assets to avoid missing built asset files in the test
environment.

* style: Disable lint warning preventing schema alignment

The `schema` key was added for alignment with existing API endpoints.

* test: Verify editor assets endpoint with lowest permission level

The `contributor` role is the lowest possible role for accessing this
endpoint, so it is worth testing with this role.

* refactor: Remove unnecessary, broken duplicate definition conditional

The class name referenced was an outdated name. The check is also
unnecessary as this file should only be loaded once.