codesniffer: Add `Jetpack.PHPUnit.TestMethodCovers` sniff (#42758)
PHPUnit 11 has effectively deprecated using `@covers` on test methods,
as the annotation is deprecated in favor of attributes that can only be
used at the class level.

This new sniff will move the `@covers` directives from the methods to
the class.

The new sniff is disabled for the moment in the monorepo itself; a
followup PR will fix all the existing instances and enable the sniff.