typescript: Use `rewriteRelativeImportExtensions` (#42990)
TypeScript used to require that imports of `.ts` files use a `.js`
extension, since that's what `tsc` would output. We have hacks in our
webpack and jest configs to account for this.

TypeScript 5.7 added a new option, `rewriteRelativeImportExtensions`,
which will rewrite extensions in relative imports (i.e. imports that
begin with `./` or `../`).

Let's switch to that option and remove the hacks. Specifically:

* Set `rewriteRelativeImportExtensions` true in
  `tools/js-tools/jest/jest-resolver.js`.
* Remove `extensionAliases` in js-packages/webpack-config.
* Remove `extensionAliases` in js-packages/storybook.
* Remove logic from `tools/js-tools/jest/jest-resolver.js`
* Update extensions in imports in JS and TS files all over the monorepo.
  And one in a webpack config.
* Fix a handful of path errors that turned up when fixing those imports.