ci: Fix CI isses after WP 6.8 release (#43090)
* Reduce CPU usage in tests by reducing password hashing cost

WordPress 6.8 switched to using PHP's `password_hash()` function, which
has a built-in default value for the 'cost' parameter which turns out to
be fairly expensive.

We don't really need the temporary users, often not even saved to a
database, to have strongly hashed passwords. So turn the cost way down
to speed up the tests.

* Use new set_transient when using WP 6.8+

Core deprecated setted_transient in favor of set_transient.

---------

Co-authored-by: tbradsha <32492176+tbradsha@users.noreply.github.com>