Blocks: do not display block placeholder to non-admins (#40209)
* Blocks: do not display block placeholder to non-admins

We currently display a placeholder in the block editor when a block cannot be used because its associated module is not active. The placeholder offers a button inviting folks to activate the module.

Since only admins can activate modules, the placeholder is useless to non-admins. Let's consequently skip registering the blocks for non-admins when the associated module is not active.

* Update placeholder status for the contact form block as well

* Add logic for VideoPress as well

* Update contact form block to handle child blocks as well

* Update logic for VideoPress package logic too

* Remove debugging

* Update lock file

* Add filter to allow third-parties to shortcircuit setup

See https://github.com/Automattic/jetpack/pull/40209/files#r1922757937

* SEO block: ensure we check for all registrations

---------

Co-authored-by: Brandon Kraft <public@brandonkraft.com>