CI: Improve selection of E2E tests to run (#43113)
If anything in `tests/e2e-commons/` changes, we should run all the E2E
tests. We do this by adding an additional `infrastructureFileSets` entry
in `jetpack dependencies` for use when the E2E setup calls it.

If `projects/plugins/jetpack/tests/e2e/specs/sync/` changes, we should
run the "Jetpack sync" suite. We do this by adding an idea of
`extraFileSets` in `jetpack dependencies`, which adds additional
projects to the output if changed files match certain regexes.