My Jetpack: Fix TS type checking and the corresponding errors (#43205)
* Add typecheck script to my-jetpack to open the can of worms

* Clean up the bad module augmentation

* Move JP_CONNECTION_INITIAL_STATE  augmentation to the connection package

The connection package should do its own job instead of other packages doing it for the package, resulting in duplication.

* subscribers-dashboard/global.d.ts file should not exist

* Let my-jetpack use the global types

When using `files` in tsconfig, it overrides the `files` of the config that it extends. Thus, moving "./global.d.ts" to `include` ensures that parent config is applied correctly

* Fix types for components package

Ideally, the components package should be standalone and portable, unaware of the logic of other packages, but unfortunately, that is not the case with these utilities. They incorporate the logic from other packages, which we try to fix by wrongly polluting the global TS namespace. Ultimately, we should get rid of these utilities from this package and move to some other package like shared extensions util

* Add style prop to text that is used in my-jetpack

* Fix JSDoc for useProductCheckoutWorkflow which breaks types

* The return type of GoldenTokenModal should be ReactNode not Component

* Fix types for WelcomeTourImage component

* Fix types for BoostCard component

* Replace __() with dummy param with _x() to make TS happy.

To avoid bad minification, we often add a third dummy param to `__()` but TS doesn't like that because the function accepts only two arguments. So, a better approach is to use `_x()` where you can pass the second argument as the context which has a benefit of giving context for the translators and also makes TS happy.

* Fix type errors caused by bad TS implementation of useRedBubbleQuery

I don't like using `as` (type assertion) but here the usage is complicated which I didn't want to play with for now.

* Add changelog

* Improve deprecation message for getSiteAdminUrl