charts: Fix tests on hosts with a negative UTC offset (#43321)
`new Date( '2024-01-03' )` creates an object for 2024-01-03T00:00:00Z.
If the host is configured with a timezone that is a negative offset from
UTC, later processing of this object interprets this as a date in the
evening of January 2 rather than the morning of January 3, causing two
tests to fail.

We can set the `TZ` environment variable to UTC to force the "local"
timezone for the tests as UTC to avoid this problem.