Use `sass-embedded` rather than `sass` (#43558)
When configured with `sass-loader` and `api: 'modern-compiler'`, it's
supposed to be faster. In some quick testing I am seeing a few seconds
saved, but most of our build time doesn't seem to be related to sass.

js-packages/charts, js-packages/image-guide, and plugins/inspect need to
be done a little hackily to avoid breaking unmaintained rollup plugins
that don't know about `sass-embedded`. Fortunately it's drop-in
compatible.

This PR also adds a pnpmfile hack to complain if something tries to
bring `sass` back in, since it would be easy for a new package to
accidentally do that.