Forms: add tracking pixel to form submission emails (#43629)
* Add tracking pixel to email form submission

Add tracking pixel to the form submission emails. To comply with GDPR,
we're not tracking any user information, just the email open email and a
hash of the admin email.

* changelog

* Fix static analysis error

* Address PR comments

I added widgth and height attributes to the img that contains the
tracking pixel to ensure the tracking pixel image stays at 1px
width/height even when it doesn't load due to DNS or some plugin
blocking it from loading.

I also added some clarifying comments regarding the use of the hash
function and the secret key.