AI Assistant: Add chrome ai token script (#43682)
* Change how we inject Chrome AI tokens

I'm changing how we load Chrome AI Origin Trial tokens to being able to
use Chrome's bult-in AI API from any domain. To do so, we have to inject
the tokens that we registered for the Origin Trials from a domain that
matches the one we used in the registration form (in this case,
widgets.wp.com).

* changelog

* Bust jetpack-chrome-ai script cache twice a day

* Fix static analysis issue